Explicitly cast to unsigned to fix a warning.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16483 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation2/components/CCmpRangeManager.cpp
@@ -1703,7 +1703,7 @@ public:
 		u16 oldMask = m_SharedDirtyVisibilityMasks[player];
 		m_SharedDirtyVisibilityMasks[player] = CalcSharedDirtyVisibilityMask(players);
 
-		if (oldMask != m_SharedDirtyVisibilityMasks[player] && player <= m_GlobalPlayerVisibilityUpdate.size())
+		if (oldMask != m_SharedDirtyVisibilityMasks[player] && (size_t)player <= m_GlobalPlayerVisibilityUpdate.size())
 			m_GlobalPlayerVisibilityUpdate[player-1] = 1;
 	}
